# src/main/resources/static/swagger/index.json
{
	"swagger": "2.0",
	"info": {
		"description": "openapi",
		"version": "3.0.0",
		"title": "益友会",
		"termsOfService": "https://www.yestae.com/home"
	},
	"host": "openapi-test.yestae.com",
	"basePath": "/api",
...
	"paths": {
...
		"/user/getAuthInfo": {
			"post": {
				"tags": ["user-api-query-controller"],
				"summary": "查询实名信息",
				"description": "查询实名信息",
				"operationId": "getAuthInfoUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uid\":\"1128947923510472705\",\"accessToken\":\"123\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/user/getInfo": {
			"post": {
				"tags": ["user-api-query-controller"],
				"summary": "查询会员信息",
				"description": "查询会员信息",
				"operationId": "getInfoUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uid\":\"1128947923510472705\",\"mobile\":\"18311112226\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/user/getInfoBatch": {
			"post": {
				"tags": ["user-api-query-controller"],
				"summary": "批量查询会员信息",
				"description": "批量查询会员信息",
				"operationId": "getInfoBatchUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uids\":[951375525788782800,951375525788782800]}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/user/orderList": {
			"post": {
				"tags": ["user-api-query-controller"],
				"summary": "查询会员历史订单",
				"description": "查询会员历史订单",
				"operationId": "orderListUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uid\":\"1128947923510472705\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/user/serviceList": {
			"post": {
				"tags": ["user-api-query-controller"],
				"summary": "查询会员尊享服务",
				"description": "查询会员尊享服务",
				"operationId": "serviceListUsingPOST",
				"consumes": ["application/json"],
				"produces": ["application/json", "*/*"],
				"parameters": [{
					"name": "appid",
					"in": "formData",
					"description": "应用编号",
					"required": true,
					"type": "string"
				}, {
					"name": "mchid",
					"in": "formData",
					"description": "商户编号",
					"required": true,
					"type": "string"
				}, {
					"name": "timestamp",
					"in": "formData",
					"description": "请求时间戳",
					"required": true,
					"type": "string"
				}, {
					"name": "version",
					"in": "formData",
					"description": "版本号",
					"required": true,
					"type": "string"
				}, {
					"name": "sign",
					"in": "formData",
					"description": "签名",
					"required": true,
					"type": "string"
				}, {
					"name": "bizcontent",
					"in": "formData",
					"description": "{\"uid\":\"1128947923510472705\"}",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
...
	},
	"securityDefinitions": {
		"token": {
			"type": "apiKey",
			"name": "token",
			"in": "header"
		}
	}
}